feature: created a fieldRobot class
