+++ core-methods/src/test/java/com/mercadopago/sdk/android/coremethods/ui/components/textfield/pcitextfield/PCITextFieldRobot.kt
@@ -2,21 +2,13 @@ package com.mercadopago.sdk.android.coremethods.ui.components.textfield.pcitextf
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.test.assert
-import androidx.compose.ui.test.hasText
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import androidx.compose.ui.test.junit4.StateRestorationTester
-import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.onNodeWithText
-import androidx.compose.ui.test.performClick
-import androidx.compose.ui.test.performTextInput
-import junit.framework.TestCase.assertEquals
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.FieldRobot
 
 internal open class PCITextFieldRobot(
-    private val composeRule: ComposeContentTestRule
-) {
-
-    internal lateinit var fieldState: PCIFieldState
+    composeRule: ComposeContentTestRule
+) : FieldRobot(composeRule) {
 
     fun createTextField(
         enabled: Boolean = true,
@@ -41,39 +33,6 @@ internal open class PCITextFieldRobot(
                 )
             }
         }
-    }
-
-    internal fun setContent(stateRestorationTester: StateRestorationTester?, content: @Composable () -> Unit) {
-        if (stateRestorationTester != null) {
-            stateRestorationTester.setContent(content)
-        } else {
-            composeRule.setContent(content)
-        }
-    }
-
-    fun performTapOnInput() {
-        composeRule.onNodeWithTag(PCITextFieldTestTags.Field.tag)
-            .assertExists()
-            .performClick()
-    }
-
-    fun performTextInput(text: String) {
-        composeRule.onNodeWithTag(PCITextFieldTestTags.Field.tag)
-            .assertExists()
-            .performTextInput(text)
-    }
-
-    fun assertTextInput(text: String) {
-        composeRule.onNodeWithTag(PCITextFieldTestTags.Field.tag)
-            .assertExists()
-            .assert(hasText(text))
-        assertEquals(text, fieldState.input)
-    }
-
-    fun assertTextIsDisplayed(text: String) {
-        composeRule.onNodeWithText(text)
-            .assertExists()
-            .assert(hasText(text))
     }
 }
