+++ core-methods/src/test/java/com/mercadopago/sdk/android/coremethods/ui/components/textfield/securityfield/SecurityFieldRobot.kt
@@ -1,15 +1,17 @@
-package com.mercadopago.sdk.android.coremethods.ui.components.textfield.pcitextfield
+package com.mercadopago.sdk.android.coremethods.ui.components.textfield.securityfield
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.test.junit4.ComposeContentTestRule
 import androidx.compose.ui.test.junit4.StateRestorationTester
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.FieldRobot
+import com.mercadopago.sdk.android.coremethods.ui.components.textfield.pcitextfield.rememberPCIFieldState
 import com.mercadopago.sdk.android.coremethods.ui.components.textfield.securitycode.SecurityCode
 import com.mercadopago.sdk.android.coremethods.ui.components.textfield.securitycode.SecurityCodeFieldEvent
 
 internal class SecurityFieldRobot(
     composeRule: ComposeContentTestRule
-) : PCITextFieldRobot(composeRule) {
+) : FieldRobot(composeRule) {
 
     fun createSecurityField(
         enabled: Boolean = true,
